AI Category and Tag Generator/backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
anthropic==0.28.0
supabase==2.5.3
python-dotenv==1.0.1
